+++ shardingsphere-jdbc/shardingsphere-jdbc-core/src/main/java/org/apache/shardingsphere/driver/jdbc/adapter/AbstractStatementAdapter.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.driver.jdbc.adapter;
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.driver.jdbc.adapter.executor.ForceExecuteTemplate;
 import org.apache.shardingsphere.driver.jdbc.unsupported.AbstractUnsupportedOperationStatement;
-import org.apache.shardingsphere.infra.executor.sql.optimize.execute.CalciteExecutor;
+import org.apache.shardingsphere.infra.executor.sql.federate.execute.FederateExecutor;
 
 import java.sql.SQLException;
 import java.sql.SQLWarning;
@@ -51,14 +51,14 @@ public abstract class AbstractStatementAdapter extends AbstractUnsupportedOperat
         closed = true;
         try {
             forceExecuteTemplate.execute((Collection) getRoutedStatements(), Statement::close);
-            closeCalciteExecutor();
+            closeFederateExecutor();
         } finally {
             getRoutedStatements().clear();
         }
     }
     
-    private void closeCalciteExecutor() throws SQLException {
-        CalciteExecutor executor = getCalciteExecutor();
+    private void closeFederateExecutor() throws SQLException {
+        FederateExecutor executor = getFederateExecutor();
         if (null != executor) {
             executor.close();
         }
@@ -212,5 +212,5 @@ public abstract class AbstractStatementAdapter extends AbstractUnsupportedOperat
     
     protected abstract Collection<? extends Statement> getRoutedStatements();
     
-    protected abstract CalciteExecutor getCalciteExecutor();
+    protected abstract FederateExecutor getFederateExecutor();
 }
